Move listing columns code into a trait
