+++ src/Http/Controllers/ResourceController.php
@@ -8,17 +8,17 @@ use DoubleThreeDigital\Runway\Http\Requests\EditRequest;
 use DoubleThreeDigital\Runway\Http\Requests\IndexRequest;
 use DoubleThreeDigital\Runway\Http\Requests\StoreRequest;
 use DoubleThreeDigital\Runway\Http\Requests\UpdateRequest;
-use DoubleThreeDigital\Runway\Resource;
 use DoubleThreeDigital\Runway\Runway;
 use DoubleThreeDigital\Runway\Support\Json;
 use Statamic\CP\Breadcrumbs;
 use Statamic\Facades\Scope;
-use Statamic\Facades\User;
 use Statamic\Fields\Field;
 use Statamic\Http\Controllers\CP\CpController;
 
 class ResourceController extends CpController
 {
+    use Traits\HasListingColumns;
+
     public function index(IndexRequest $request, $resourceHandle)
     {
         $resource = Runway::findResource($resourceHandle);
@@ -242,32 +242,6 @@ class ResourceController extends CpController
         return [
             'data' => $this->getReturnData($resource, $record),
         ];
-    }
-
-    /**
-     * This method is a duplicate of code in the `ResourceListingController`.
-     * Update both if you make any changes.
-     */
-    protected function buildColumns(Resource $resource, $blueprint)
-    {
-        $preferredFirstColumn = isset(User::current()->preferences()['runway'][$resource->handle()]['columns'])
-            ? User::current()->preferences()['runway'][$resource->handle()]['columns'][0]
-            : $resource->listableColumns()[0];
-
-        return collect($resource->listableColumns())
-            ->map(function ($columnKey) use ($blueprint, $preferredFirstColumn) {
-                $field = $blueprint->field($columnKey);
-
-                return [
-                    'handle' => $columnKey,
-                    'title' => $field
-                        ? $field->display()
-                        : $field,
-                    'has_link' => $preferredFirstColumn === $columnKey,
-                    'is_primary_column' => $preferredFirstColumn === $columnKey,
-                ];
-            })
-            ->toArray();
     }
 
     /**
