+++ src/Http/Controllers/ResourceListingController.php
@@ -3,7 +3,6 @@
 namespace DoubleThreeDigital\Runway\Http\Controllers;
 
 use DoubleThreeDigital\Runway\Http\Resources\ResourceCollection;
-use DoubleThreeDigital\Runway\Resource;
 use DoubleThreeDigital\Runway\Runway;
 use Statamic\Facades\User;
 use Statamic\Http\Controllers\CP\CpController;
@@ -12,7 +11,7 @@ use Statamic\Query\Scopes\Filters\Concerns\QueriesFilters;
 
 class ResourceListingController extends CpController
 {
-    use QueriesFilters;
+    use QueriesFilters, Traits\HasListingColumns;
 
     public function index(FilteredRequest $request, $resourceHandle)
     {
@@ -68,31 +67,5 @@ class ResourceListingController extends CpController
             ->additional(['meta' => [
                 'activeFilterBadges' => $activeFilterBadges,
             ]]);
-    }
-
-    /**
-     * This method is a duplicate of code in the `ResourceController`.
-     * Update both if you make any changes.
-     */
-    protected function buildColumns(Resource $resource, $blueprint)
-    {
-        $preferredFirstColumn = isset(User::current()->preferences()['runway'][$resource->handle()]['columns'])
-            ? User::current()->preferences()['runway'][$resource->handle()]['columns'][0]
-            : $resource->listableColumns()[0];
-
-        return collect($resource->listableColumns())
-            ->map(function ($columnKey) use ($blueprint, $preferredFirstColumn) {
-                $field = $blueprint->field($columnKey);
-
-                return [
-                    'handle' => $columnKey,
-                    'title'  => $field
-                        ? $field->display()
-                        : $field,
-                    'has_link' => $preferredFirstColumn === $columnKey,
-                    'is_primary_column' => $preferredFirstColumn === $columnKey,
-                ];
-            })
-            ->toArray();
     }
 }
